requirements.txt
streamlit>=1.36.0
openai>=1.37.0
transformers>=4.42.0
torch>=2.3.0
accelerate>=0.31.0
llama-cpp-python>=0.2.90
